[MOL-22129][XF] fix link list e2e test
