+++ e2e/tests/components/link-list/link-list.e2e.spec.ts
@@ -81,6 +81,16 @@ test.describe("LinkList", () => {
         });
     });
 
+    test.describe(() => {
+        test.beforeEach(async ({ story }) => {
+            await story.init("custom-content");
+        });
+
+        test("Custom content", async ({ story }) => {
+            await compareScreenshot(story, "mount");
+        });
+    });
+
     test.describe("Eager load", () => {
         test.describe(() => {
             modes.forEach((mode) => {
@@ -169,24 +179,6 @@ test.describe("LinkList", () => {
                     ).toBeFocused();
                 });
             });
-        });
-    });
-
-    test.describe(() => {
-        test.beforeEach(async ({ story }) => {
-            await story.init("jsx-content");
-        });
-
-        test("JSX title and description", async ({ story }) => {
-            const jsxTitleLink = story.locators.link("JSX title item");
-            await expect(jsxTitleLink.locator("strong")).toHaveText("title");
-
-            const jsxDescLink = story.locators.link(
-                "Item with JSX description"
-            );
-            await expect(jsxDescLink.locator("strong")).toHaveText("rich");
-
-            await compareScreenshot(story, "mount");
         });
     });
